Add support for complex input variables [WIP]
As requested in #7

This is incomplete, but it already works with primitive types and string
lists.

Needs to be extended to support complex list types.

Maps and objects may be added as well, but first lists need to be
implemented fully.

Also, the old state needs to be migrated to the new one (Variables type
changed) – or maybe let's introduce a new name and keep the old one for
compatibility.